requirements.txt
requests==2.32.3 
alpaca-py==0.28.1 
datetime==5.5 
pandas~=2.2.3
numpy~=1.26.4
statsmodels==0.14.2
matplotlib~=3.9.1.post1
